Add files via upload
File Amazon-product-scraper.py contains the Python program to scrape atleast 20 pages of product listing pages. And the file products-part1-assignment.csv contains the output of the first part of the assignment.
File Product-URL-Hit-part 2.py contains the Python program to hit product URL and to fetch various required information. And the file asignment-part2-products-url-hit.csv
contains the output of the second part of the assignment.